Clean can truncate instead of drop and create tables

+++ Fixtures/FixtureManager.php
@@ -359,12 +359,17 @@ class FixtureManager implements FixtureManagerInterface
         }
     }
 
-    public function clean()
+    public function clean($softReset = false)
     {
         $this->classes = array_unique($this->classes);
 
         $schemaTool = $this->getSchemaTool();
-        $schemaTool->dropSchema($this->classes);
-        $schemaTool->createSchema($this->classes);
+
+        if ($softReset) {
+            $schemaTool->truncateSchema($this->classes);
+        } else {
+            $schemaTool->dropSchema($this->classes);
+            $schemaTool->createSchema($this->classes);
+        }
     }
 }
